(#5) Repositorios. Se crea repositorio base para los métodos de propiedades más comunes. Se mueve método findByType a AppRepository.
