+++ src/Repository/AppErrorRepository.php
@@ -3,10 +3,8 @@
 namespace App\Repository;
 
 use App\Entity\AppError;
-use Doctrine\ORM\NonUniqueResultException;
 use Doctrine\Persistence\ManagerRegistry;
 use App\Repository\Interfaces\AppErrorRepositoryInterface;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 
 /**
  * @method AppError|null find($id, $lockMode = null, $lockVersion = null)
@@ -14,13 +12,13 @@ use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
  * @method AppError[]    findAll()
  * @method AppError[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class AppErrorRepository extends ServiceEntityRepository implements AppErrorRepositoryInterface
+class AppErrorRepository extends AppRepository implements AppErrorRepositoryInterface
 {
 
     /************************************************* CONSTRUCT **************************************************/
 
     /**
-     * AppError constructor.
+     * AppErrorRepository constructor.
      *
      * @param ManagerRegistry $registry Object managers for a Doctrine persistence layer ManagerRegistry class.
      */
@@ -30,22 +28,6 @@ class AppErrorRepository extends ServiceEntityRepository implements AppErrorRepo
     }
 
     /*********************************************** PUBLIC METHODS ***********************************************/
-
-    /**
-     * @inheritDoc
-     * @return array array
-     */
-    public function findByType(int $type): array
-    {
-        $alias = 'err';
-
-        return $this->createQueryBuilder($alias)
-            ->andWhere($alias . '.type = :type')
-            ->setParameter('type', $type)
-            ->orderBy($alias . '.id', 'ASC')
-            ->getQuery()
-            ->getArrayResult();
-    }
 
     /*********************************************** STATIC METHODS ***********************************************/
